tempo env var faster timeout for when if http end point is down
